# openapi.yaml
openapi: 3.0.3
info:
  title: Unico Napp API
  version: 1.0.0
  description: |
    Documentação OpenAPI da API Unico Napp.
    Cada recurso contém explicações e exemplos de requisição/ resposta.
servers:
  - url: http://localhost:5000
    description: Servidor local de desenvolvimento
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    SimpleErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid API key
    ErrorItem:
      type: object
      properties:
        codigo_erro:
          type: integer
          example: 1001
        descricao:
          type: string
          example: username - Possui valor inesperado ou vazio
    ErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
  responses:
    AdminKeyRequired:
      description: Admin key is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SimpleErrorResponse'
          example:
            error: Admin key required for registration
    UsernameRequired:
      description: Username is required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SimpleErrorResponse'
          example:
            error: username required
    UsernameExists:
      description: Username already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SimpleErrorResponse'
          example:
            error: username exists
...
    InvalidApiKey:
      description: API key is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SimpleErrorResponse'
          example:
            error: Invalid API key
...
    ProductNotFound:
      description: Product was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - codigo_erro: 1013
                descricao: id_produto - produto não encontrado
...
    ImageNotFound:
      description: Image was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - codigo_erro: 1022
                descricao: Recurso não encontrado
...
    WebhookValidationError:
      description: Webhook URL is missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SimpleErrorResponse'
          example:
            error: webhook_url is required
...
paths:
  /api/v1/auth/register:
    post:
      summary: Registrar usuário
      description: Registra um novo usuário e gera uma API key para uso nas integrações.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: Usuário registrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRegisterResponse'
        '400':
          $ref: '#/components/responses/UsernameRequired'
        '403':
          $ref: '#/components/responses/AdminKeyRequired'
        '409':
          $ref: '#/components/responses/UsernameExists'
  /api/v1/auth/me:
    get:
      summary: Informações do usuário
      description: Retorna informações do usuário autenticado e a URL de webhook configurada.
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Dados do usuário
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '401':
          $ref: '#/components/responses/InvalidApiKey'
  /api/v1/auth/webhook:
    post:
      summary: Configurar webhook do usuário
      description: Salva ou atualiza a URL que receberá eventos de status de pedido.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookConfigRequest'
      responses:
        '200':
          description: Webhook configurado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookConfigResponse'
        '400':
          $ref: '#/components/responses/WebhookValidationError'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SimpleErrorResponse'
              examples:
                missing_api_key:
                  value:
                    error: API key required
                invalid_api_key:
                  value:
                    error: Invalid API key
...
  /api/v1/imagens:
    get:
      summary: Listar imagens
      description: Lista imagens de produtos com paginação e filtro opcional por produto.
      parameters:
        - name: produto_id
          in: query
          schema:
            type: integer
          description: ID do produto para filtrar imagens
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de imagens
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Image'
    post:
      summary: Criar imagem
      description: Adiciona uma imagem a um produto existente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageCreateRequest'
      responses:
        '201':
          description: Imagem criada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  message:
                    type: string
                    example: created
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  value:
                    errors:
                      - codigo_erro: 1001
                        descricao: link - Possui valor inesperado ou vazio
                product_not_found:
                  value:
                    errors:
                      - codigo_erro: 1013
                        descricao: id_produto - produto não encontrado
  /api/v1/imagens/{id}:
    get:
      summary: Consultar imagem por ID
      description: Retorna os dados de uma imagem pelo seu identificador.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Imagem encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
        '404':
          $ref: '#/components/responses/ImageNotFound'
    put:
      summary: Atualizar imagem
      description: Substitui todos os dados de uma imagem existente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageCreateRequest'
      responses:
        '200':
          description: Imagem atualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: updated
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  value:
                    errors:
                      - codigo_erro: 1001
                        descricao: ordem - Possui valor inesperado ou vazio
                product_not_found:
                  value:
                    errors:
                      - codigo_erro: 1013
                        descricao: id_produto - produto não encontrado
        '404':
          $ref: '#/components/responses/ImageNotFound'
    patch:
      summary: Atualizar parcialmente imagem
      description: Atualiza apenas os campos fornecidos da imagem.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Imagem atualizada parcialmente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: updated
        '400':
          $ref: '#/components/responses/ProductNotFound'
        '404':
          $ref: '#/components/responses/ImageNotFound'
  /api/v1/imagens/produto/{produto_id}:
    get:
      summary: Listar imagens por produto
      description: Retorna todas as imagens vinculadas a um produto específico.
      parameters:
        - name: produto_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de imagens do produto
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Image'
...